feat: implement TODO items, move tests to tests/, add rustdoc CI
- Implement `for_each_integer_type!` macro (resolves TODO "Create macro foreach")
- Document IDE alias design decision in LinkReference (resolves TODO "Not use alias")
- Move all tests from src/imp.rs to tests/traits.rs (integration tests)
- Add 8 new tests (ToPrimitive, PrimInt ops, FromPrimitive, Send+Sync, Display, Hash)
- Add crate-level documentation in lib.rs with trait summary table
- Add deploy-rust-docs.yml workflow for automated docs on GitHub Pages
- Update case study with new findings
