+++ .github/workflows/deploy-rust-docs.yml
@@ -1,0 +1,34 @@
+name: Deploy Rust Docs
+
+on:
+  push:
+    branches: main
+    paths:
+      - 'rust/**'
+      - '.github/workflows/deploy-rust-docs.yml'
+  workflow_dispatch:
+
+permissions:
+  contents: write
+
+jobs:
+  deploy-docs:
+    name: Build and Deploy Rust Documentation
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Setup Rust
+        uses: dtolnay/rust-toolchain@stable
+
+      - name: Build documentation
+        working-directory: rust
+        run: cargo doc --no-deps
+
+      - name: Deploy to GitHub Pages
+        uses: peaceiris/actions-gh-pages@v4
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          publish_dir: rust/target/doc
+          destination_dir: rust
+          keep_files: true
