+++ docs/case-studies/issue-142/README.md
@@ -52,11 +52,15 @@ The issue requested a comprehensive audit of the Rust codebase:
 
 - **Clippy:** Zero warnings with all targets and features enabled.
 - **Formatting:** Fully compliant with `rustfmt`.
-- **Tests:** 67 unit tests + 5 doc tests, all passing. 100% code coverage.
+- **Tests:** 75 integration tests + 6 doc tests, all passing.
 - **Architecture:** Clean trait design with blanket implementations and zero-cost abstractions.
 
-**Minor issues found and fixed:**
-- **Stale TODO comments:** Two TODO comments in `imp.rs` (lines 117 and 151) referenced work that was either already addressed or no longer relevant. Removed.
+**Improvements made:**
+
+- **Implemented `for_each_integer_type!` macro:** The original TODO comment `// TODO: Create macro foreach` requested a macro to DRY up repetitive per-type macro invocations. Implemented the `for_each_integer_type!` helper macro that invokes a given macro for all 12 primitive integer types.
+- **Documented IDE alias design decision:** The original TODO comment `// TODO: Not use alias - IDEs does not support it` was about the `LinkReference` trait listing supertraits explicitly. Added a `# Design note` in the doc comment explaining why aliases are not used (IDE compatibility).
+- **Moved tests to `tests/` directory:** All unit tests were in `src/imp.rs`. Moved them to `rust/tests/traits.rs` as proper integration tests, following Rust best practices for separation of tests from source code.
+- **Added new tests:** Added 8 additional integration tests covering `ToPrimitive`, `PrimInt` operations, `FromPrimitive`, `Send + Sync`, `Display`, `Hash`, and overflow scenarios. Total: 75 integration tests + 6 doc tests = 81 tests.
 - **Cargo.lock version mismatch:** Lock file showed version `0.4.0` while `Cargo.toml` was at `0.5.0`. Regenerated.
 
 ### 5. Documentation Sync
@@ -67,26 +71,38 @@ The issue requested a comprehensive audit of the Rust codebase:
 |----------|-------|-----|
 | `rust/README.md` | Installation example showed `platform-num = "0.4"` | Updated to `"0.5"` |
 | `CONTRIBUTING.md` | Stated project uses "a specific nightly toolchain" | Corrected to "the stable toolchain" |
-| `rust/README.md` | Trait documentation | Already in sync with code — no changes needed |
-| Doc comments in `imp.rs` | All five traits have `///` docs with examples | Already in sync — no changes needed |
+| `rust/src/lib.rs` | No crate-level documentation | Added comprehensive module-level docs with trait summary table and example |
+| `rust/src/imp.rs` | All five traits have `///` docs with examples | Already in sync — enhanced `LinkReference` with design note |
+
+### 6. Automated Documentation Generation
+
+**Finding:** The C# part of the project already has automated docs via `docfx`, deployed to GitHub Pages at `csharp/` subdirectory on the `gh-pages` branch.
+
+**Action taken:** Added `.github/workflows/deploy-rust-docs.yml` workflow that:
+- Triggers on push to `main` when Rust files change
+- Runs `cargo doc --no-deps` to generate HTML documentation
+- Deploys to `rust/` subdirectory on the `gh-pages` branch
+- Uses `peaceiris/actions-gh-pages@v4` (same approach as the C# docs)
+
+This makes Rust documentation available alongside C# documentation on the project's GitHub Pages site.
 
 ---
 
 ## Solution Plan
 
-### Approach: Minimal, Targeted Changes
+### Approach: Targeted Improvements
 
-Since the codebase was already in good shape, the approach was to make only the changes that genuinely improve quality:
+1. **Edition upgrade (2021 → 2024):** Adopts the latest stable edition, bringing the crate in line with modern Rust ecosystem standards. Zero code changes required.
 
-1. **Edition upgrade (2021 → 2024):** The most impactful change. Adopts the latest stable edition, bringing the crate in line with modern Rust ecosystem standards. Zero code changes required — the migration was seamless.
+2. **MSRV bump (1.70 → 1.85):** Required by Edition 2024. The MSRV-aware resolver (stabilized in Rust 1.84) ensures downstream users on older toolchains automatically get compatible older versions.
 
-2. **MSRV bump (1.70 → 1.85):** Required by Edition 2024 (minimum Rust 1.85). This is reasonable because the MSRV-aware resolver (stabilized in Rust 1.84) ensures downstream users on older toolchains automatically get compatible older versions of this crate.
+3. **Documentation improvements:** Fixed factual inaccuracies, added crate-level docs, enhanced trait documentation with design notes.
 
-3. **Documentation corrections:** Fixed factual inaccuracies (nightly vs stable, version 0.4 vs 0.5) that could mislead contributors and users.
+4. **TODO resolution:** Instead of removing TODOs, implemented the requested `for_each_integer_type!` macro and documented the IDE alias design decision.
 
-4. **TODO cleanup:** Removed stale TODO comments that added noise without value.
+5. **Test restructuring:** Moved tests from `src/` to `tests/` directory and added new test coverage.
 
-5. **Cargo.lock regeneration:** Ensured lock file reflects the actual published version.
+6. **Automated docs deployment:** Added CI workflow for Rust documentation generation and GitHub Pages deployment.
 
 ---
 
@@ -116,6 +132,14 @@ The following Edition 2024 changes were evaluated for impact on this crate:
 
 **Result:** Clean migration with zero code changes.
 
+### Rust Documentation Generation
+
+For Rust projects, `cargo doc` is the standard documentation generator (equivalent to C#'s `docfx`). It:
+- Parses `///` and `//!` doc comments using Markdown
+- Generates cross-linked HTML with search
+- Supports code examples that are compiled and tested (`cargo test --doc`)
+- Is built into the Rust toolchain — no additional dependencies needed
+
 ---
 
 ## Verification
@@ -124,6 +148,6 @@ All quality checks pass after changes:
 
 - `cargo fmt --check` — ✅ clean
 - `cargo clippy --all-targets --all-features` — ✅ zero warnings
-- `cargo test --all-features --verbose` — ✅ 67 unit tests + 5 doc tests pass
-- `cargo test --doc --verbose` — ✅ all doc tests pass
+- `cargo test --all-features` — ✅ 75 integration tests + 6 doc tests pass
+- `cargo doc --no-deps` — ✅ documentation generates successfully
 - Edition 2024 compilation — ✅ no migration issues
